+++ R/find_weights.R
@@ -63,3 +63,24 @@ find_weights.brmsfit <- function(x, ...) {
   if (.is_empty_object(w)) w <- NULL
   w
 }
+
+
+#' @export
+find_weights.merMod <- function(x, ...) {
+  tryCatch(
+    {
+      w <- .safe_deparse(parse(text = .safe_deparse(x@call))[[1]]$weights)
+
+      # edge case, users use "eval(parse())" to parse weight variables
+      if (grepl("^eval\\(parse\\(", w)) {
+        w <- eval(parse(text = .trim(gsub("eval\\(parse\\((.*)=(.*)\\)\\)", "\\2", w))))
+      }
+
+      if (.is_empty_object(w) || w == "NULL") w <- NULL
+      w
+    },
+    error = function(e) {
+      NULL
+    }
+  )
+}